Рефакторинг. Избавится от моков фильтров и перенесет логику их работы в компонент фильтров.
