+++ src/components/filters.js
@@ -1,18 +1,55 @@
 import AbstractSmartComponent from "./abstract-smart-component.js";
 
-const createFilterMarkup = (filter, films) => {
-  const {caption, hasCounter, link, isActive, action} = filter;
-  const counterMarkup = hasCounter ? ` <span class="main-navigation__item-count">${action(films)}</span>` : ``;
 
-  return `<a href="${link}" class="main-navigation__item ${isActive ? `main-navigation__item--active` : ``}">${caption}${counterMarkup}</a>`;
+const FiltersData = {
+  "ALL_MOVIES": {
+    DATA_TAG: `all-movies`,
+    CAPTION: `All movies`,
+    HAS_COUNTER: false,
+    LINK: `#all`,
+    IS_ACTIVE: true,
+    ACTION: (films) => films,
+  },
+  "WATCHLIST": {
+    DATA_TAG: `watchlisted`,
+    CAPTION: `Watchlist`,
+    HAS_COUNTER: true,
+    LINK: `#watchlist`,
+    IS_ACTIVE: false,
+    ACTION: (films) => films.reduce((total, it) => it.isWatchlisted ? total + 1 : total, 0),
+  },
+  "HISTORY": {
+    DATA_TAG: `watched`,
+    CAPTION: `History`,
+    HAS_COUNTER: true,
+    LINK: `#history`,
+    IS_ACTIVE: false,
+    ACTION: (films) => films.reduce((total, it) => it.isWatched ? total + 1 : total, 0),
+  },
+  "FAVORITES": {
+    DATA_TAG: `favorites`,
+    CAPTION: `Favorites`,
+    HAS_COUNTER: true,
+    LINK: `#favorites`,
+    IS_ACTIVE: false,
+    ACTION: (films) => films.reduce((total, it) => it.isFavorite ? total + 1 : total, 0),
+  },
 };
 
-const createFiltersMarkup = (filters, films) => {
+
+const createFilterMarkup = (filter, films) => {
+  const counterMarkup = filter.HAS_COUNTER ? ` <span class="main-navigation__item-count">${filter[`ACTION`](films)}</span>` : ``;
+  const isFilterActive = filter.IS_ACTIVE ? `main-navigation__item--active` : ``;
+
+  return `<a href="${filter.LINK}" data-filter-type=${filter.DATA_TAG} class="main-navigation__item ${isFilterActive}">${filter.CAPTION}${counterMarkup}</a>`;
+};
+
+const createFiltersMarkup = (films) => {
   let targetMarkup = ``;
 
-  for (const filterKey in filters) {
-    if (filters.hasOwnProperty(filterKey)) {
-      targetMarkup = targetMarkup + createFilterMarkup(filters[filterKey], films);
+  for (const filterKey in FiltersData) {
+    if (FiltersData.hasOwnProperty(filterKey)) {
+      targetMarkup = targetMarkup + createFilterMarkup(FiltersData[filterKey], films);
     }
   }
   return targetMarkup;
@@ -32,12 +69,11 @@ export default class Filters extends AbstractSmartComponent {
   constructor(filters, films) {
     super();
 
-    this._filters = filters;
     this._films = films;
   }
 
   getTemplate() {
-    return createFiltersTemplate(this._filters, this._films);
+    return createFiltersTemplate(this._films);
   }
 
   recoverListeners() {
@@ -48,4 +84,28 @@ export default class Filters extends AbstractSmartComponent {
     this._films = newFilms;
     super.rerender();
   }
+
+  // setFilterTypeChangeHandler(handler) {
+  //   this.getElement().addEventListener(`click`, (evt) => {
+  //     evt.preventDefault();
+  //
+  //     if (evt.target.tagName !== `A`) {
+  //       return;
+  //     }
+  //
+  //     const sortType = evt.target.dataset.sortType;
+  //
+  //     if (this._currentSortType === sortType) {
+  //       return;
+  //     }
+  //
+  //     this._currentSortType = sortType;
+  //     this.getElement().querySelector(`.${SortButtonCls.ACTIVE}`).classList.remove(SortButtonCls.ACTIVE);
+  //     evt.target.classList.add(SortButtonCls.ACTIVE);
+  //
+  //     handler(this._currentSortType);
+  //   });
+  // }
 }
+
+// export {FilterType};
