+++ src/main.js
@@ -4,7 +4,6 @@ import MainNavigationComponent from "./components/main-navigation.js";
 import PageController from "./controllers/page.js";
 import UserRankComponent from "./components/user-rank.js";
 import {generateFilms} from "./mock/films.js";
-import {generateFilters} from "./mock/filters.js";
 import {render} from "./utils/render.js";
 
 
@@ -14,7 +13,6 @@ const FILMS_ALL_COUNT = 13;
 const films = generateFilms(FILMS_ALL_COUNT);
 const filmsTotal = films.length;
 const userFilmsWatched = films.reduce((total, it) => it.isWatched ? total + 1 : total, 0);
-const filters = generateFilters();
 
 const siteHeaderElement = document.querySelector(`.header`);
 render(siteHeaderElement, new UserRankComponent(userFilmsWatched));
@@ -26,5 +24,5 @@ render(siteMainElement, siteMainNavigationComponent);
 const siteFilmsTotalElement = document.querySelector(`.footer__statistics`);
 render(siteFilmsTotalElement, new FilmsTotalComponent(filmsTotal));
 
-const pageController = new PageController(siteMainElement, films, filters);
+const pageController = new PageController(siteMainElement, films);
 pageController.render();
